http_server/requirements.txt
pydantic==1.8.1
fastapi==0.63.0
mysql-connector-python==8.0.23
